requirements.txt
openshift-client>=1.0.14
